chip/stm32/dfu_bootmanager_main: Add memory barrier
ARM Cortex documentation calls for an ISB instruction to be executed
immediately after modifying the stack pointer, see for instance:
https://developer.arm.com/documentation/dui0552/a/the-cortex-m3-processor/programmers-model/core-registers

It would seem that instruction prefetch will "anticipate" stack pointer
movements by counting the push and pop instructions, and that the
prefetch queue needs to be cleared if the stack pointer is manually
changed.  Our particular code executes does not use the stack pointer,
but executes a register jump, which I assume will also clear the
prefetch queue.  So in all likelihood, the barrier instructions is not
required here, but we could as well add it, as the documentation calls
for it.

Change-Id: I2a0609670b3554a39057b429fc2e0cd1201d185c