partner emul: Handle Accept, Reject consistently
When the partner emulator receives a Reject, consistently call Reject
handlers. When the partner receives an Accept, consistently call Accept
handlers. Do not combine Accept and Reject logic. After the partner
sends DR_Swap, avoid sending Soft_Reset in response to Accept.

Verify that the TCPM will Accept a DR_Swap when the TCPM is UFP and that
the TCPM will actually switch data roles without a disconnection.

BUG=b:262345042
TEST=twister -s zephyr/test/drivers/drivers.default
BRANCH=none

Change-Id: Iab61d912a1acd7300bedf251f4856e4c3d86abf6
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>