chip/stm32/clock-stm32l4: Fixed clock cmd documentation
Looking at the code, it is apparent that the "pll" parameter to the
"clock" console command is valid whether STM32_HSE_CLOCK is defined or
not (though the effect may be slightly different).

This CL updates the documentation to list "pll" as a valid option also
when STM32_HSE_CLOCK is not defined.

BUG=none
TEST=make BOARD=hyperdebug

Change-Id: I7da89c15b44fbc3dad2a5fe129b0e361d7f50714