core/cortex-m: Allow copying vector table into SRAM
A single STM chip family of the Cortex-M0 core routinely copies the
vector table into SRAM.  The vector table has to be aligned on a 512
byte boundary, so a special region, put very first SRAM is used.

This CL makes it such that boards using other kinds of Cortex-M can
optionally copy the vector table into SRAM, by declaring the same
special linker section.

If the boards do not use it, the section will not take up any space in
SRAM.

BUG=b:266832220

Change-Id: I6222f360f52ec4928327860fd1bf78860dfba120