pre-upload: Run check_zephyr_project_config in vpython3
All hooks scripts are expected to work inside and outside of the
chroot.  This one does not when run outside the chroot:

Hook script "util/check_zephyr_project_config.py -d ${PRESUBMIT_FILES}" failed with code 1:
    Traceback (most recent call last):
    File "/home/jrosenth/chromiumos/src/platform/ec/util/check_zephyr_project_config.py", line 36, in <module>
        import zephyr_module
    File "/home/jrosenth/chromiumos/src/third_party/zephyr/main/scripts/zephyr_module.py", line 27, in <module>
        import pykwalify.core
    ModuleNotFoundError: No module named 'pykwalify'

By executing in vpython3, and brining in the standard Zephyr
dependencies, this resolves this check's failure inside and outside
the chroot.

BUG=none
BRANCH=none
TEST='repo upload' inside & outside the chroot

Change-Id: Ied51dc3b6886c02a010e758b0ef23364d9232b99
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>