Increase supported host events to 64
ectool is hard-coded to support 32 host events even though the EC
supports 64 host events.

Add an enum value to indicate what the current count of the host events
is, so it can used by ectool in place of magic numbers.

Also add an assert to enforce the 64 host event limit.

BRANCH=none
BUG=b:261141172
TEST=Manually build and flash, verify device boots
TEST=ectool version

Change-Id: If9724cf905a7d2eb42a2ad67c5e1da784ca05e8e
Signed-off-by: Tim Van Patten <timvp@google.com>