it8xxx2: micro-optimize SHA256 computation
Hash performance is a significant contributor to EC boot-up time, so it's
valuable to optimize the computation of firmware hashes. This change
improves RW image hashing performance on Nereid, reducing time taken
for `hash rw` on the console by about 15ms on average
(from 1263ms to 1244ms for 384kB of data).

This primarily replaces the byte swapping in SHA256_update with an inline
assembly version that saves 6 instructions per word of data hashed. Calling
htobe32() turns into a call into libgcc's __bswapsi2 which does implement
an efficient swap, but spends several instructions reloading constant
values used by the function and incurs function call overhead. This inline
version saves 4 instructions by allowing the constants to be kept in
registers through the entire SHA256_update (effectively hoisting the
constant loads outside the loop), and 2 more in procedure call overhead.

SHA256_chip_calculation is also optimized slightly: declaring the
hash_ctrl local as volatile forces it to be allocated a stack slot and
the read value be written to the stack (then ignored), which is unnecessary
because the macro referring to the SHA1HASHCTRLR register is already a
volatile read. Removing the unneeded qualifier saves 3 instructions between
stack adjustment that is no longer needed (because the function now
operates only on values that fit in registers) and the unneeded store that
is now removed.

BUG=b:260762509
TEST=`hash 0xbfffc 32` has the same value before and after this change,
     with manual verification that the 32 bytes of memory at 0x800bfffc
     are unchanged between the two builds (`md 0x800bfffc 32`).
BRANCH=nissa

Change-Id: I99fcb2278518f6a57046985ef03a58ebb2f307a1
Signed-off-by: Peter Marheine <pmarheine@chromium.org>