trogdor: Do not change the TCPC MUX when HPD changes
During the DP-Config phase, when the DP-Config VDM is acked, the
TCPC MUX is connected. Afterward, when the EC receives any HPD
event (High, Low, or IRQ), should not change the TCPC MUX.
Otherwise, disconnecting the TCPC MUX when HPD Low makes the AUX
termination disappear.

BRANCH=trogdor
BUG=b:247005613
TEST=Connect the device through a Type-C dongle to a monitor,
unplug the monitor from the dongle, and replug it again. The
monitor should detect the device and assert HPD high.

Change-Id: Id890bfe8fec23109da2b36aa9662e24aec4ae32c
Signed-off-by: Wai-Hong Tam <waihong@google.com>