TCPMv1/v2: Add config option to disable TCPC VCONN
Currently we always source VCONN from the TCPC, and also the PPC if
present. However the SYV682 can't handle 5V on its host-side CC pins, so
we shouldn't source VCONN in that case.

For those TCPCs which will trigger OVP if VCONN isn't enabled, this will
not happen with the SYV682, since it isolates VCONN from the TCPC CC
signals.

BUG=b:269050063
BRANCH=none
TEST=On marasov, make sure the host-side CC pin are <3.3v when sourcing
VCONN.

Change-Id: I1f979acfc1f157f42ab0036935c477c998b1bb1f
Signed-off-by: Michael5 Chen1 <michael5_chen1@pegatron.corp-partner.google.com>