binman: Add CBI section to fmap
CBI section is added to fmap. The CBI section will be filled with 0xFF
in binary. The size and offset could be overridden with device tree
config in chip/program/project.

A property named "preserve" is added, to determine the CBI portion
should not be updated during flashing of chip.

BRANCH=None
BUG=b:266972341
TEST=dump_fmap <ec.bin location> && manually checking ec.bin content

Change-Id: I43647285502917bf557b2fb4ce174c24538fb363
Signed-off-by: Madhurima Paruchuri <mparuchuri@google.com>