board/hyperdebug: Analog input support
The HyperDebug board is a versatile debugger, serving a similar purpose
to uServo for development boards for Google Security Chips or other
microcontrollers, rather than full Chromebooks.

This change allows the ADC circuits of HyperDebug to be enabled, and
allows about a dozen pins to be used as analog input.

BUG=b:269621551

Change-Id: I437412260003712ac4a5ff27f969182612feb35e