flash_ec: delay cold_reset releasing time
Using cold_reset (which was EC reset, and now changes to GSC reset),
and not enough sleep time would cause a race condition of FW_UP_L
and EC_RST. The booting sequence would be:

1. GSC_RESET on
2. EC_RESET on
2. FW_UP_L on
3. GSC_REST off
4. EC_RESET off
5. FW_UP_L off

The race condition happens at the step 4 and step 5 if the EC reset
released too lade.

This causes an issue entering the EC flashing mode on the new Ti50
platforms. We extend the sleep time between FW_UP_L release and
COLD_RESET release to 0.2, which should be sufficient by the
data we measured.

BUG=b:269182955
TEST=flash over ccd and servo_micro on hayato (cr50,ite),
     kingler (ti50, npcx), geralt (ti50, ite)
BRANCH=none

Change-Id: I5a8b5c30bc44779f03ec4ed95cb505e7aaf1a0f0
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>