chip/stm32: ADC fixes
The ADC drivers for STM32L4 (and L5) seems to have significant flaws.

The chip has the ability to continuously perform ADC conversions
according to a set sequence of channels and bit depths.  In addition to
that, the chip can perform a one-off "injected" sequence of up to four
channels, in response to a hardware or software trigger.

The driver uses only a software triggered "injected" sequence to convert
two pre-set channels.

The code of `adc_read_channel()` has some odd properties, though.  On
first invocation, it sets up the injected sequence to consist of the two
configured ADC channels.  Each invocation takes one particular ADC
channel number as input, but the code simply kicks of the sequence of
two channels to read, and then returns the relevant of the two,
discarding the other reading.

The has the needless limitation that it cannot be used with more than
two ADC channels.

Since it is permitted to modify the list of channels in the injected
sequence as long as no conversion is in progress, it would be more
straightforward, if each invocation of `adc_read_channel()` would
reconfigure the injected sequence to consist of a single channel, the
one requested.  This CL makes that change.

BUG=b:269621551
TEST=Read and of a dozen ADC channels on HyperDebug

Change-Id: I62387979faf494cfefc3b6e7dd1d9a1954017ae6