chip/stm32: Sample time off-by-one
The STM32L4/5 ADCs use three bits to configure the time to sample the
input voltage, that is the time during which an analog MUX conducts,
allowing the internal holding capacitor to settle.  The enums declared
in adc_chip.h has eight usable values, plus an additional value called
DEFAULT.

It makes sense to have the DEFAULT be represented by a zero value, such
that if a board.c file does not mention the sample_time field in its
adc_t, then due to the C convention of data being zero-initialized, the
code can detect it and apply a sensible default.

However, the body of adc_configure() has no such logic, but takes the
raw enum value and attempts to stuff it into a 3-bit field, which will
be off by one, and overrun for the largest enum value of 8.

Also, I find the term "sample rate" misleading, as it implies a
continuous process.  This sample time setting applies equally to one-off
injected conversions, as it refers to the duration of the sampling of
the input signal, before conversion begins, not a rate of conversions
happening.

BUG=b:269621551
TEST=todo

Change-Id: Id8d297fcec883565dea1e09d6bbbfa1ab564778d