chip/stm32: Add warnings in stm32l4 files also used by L5
In several cases, the L4 and L5 series can use the same driver, which is
different from other STM32Lx chips.

This CL introduce a warning at the top of such files, alerting readers
to the fact that they are also used by L5 despite this not being
apparent from the file name.

BUG=b:269621551
TEST=none

Change-Id: I6edff6c0aea57ea9679729db20a292d75b4b8df0