ppc/syv682x: Handle incorrect Vconn config
As the maximum voltage of HOST_CCx pinx on syv682x is 3.6V, the Vconn
sourced by TCPC may exceed 3.6V and damage syv682x.
The syv682x init function checks the configuration of Vconn. The init
function should fail with an error message if TCPC is configured to
control Vconn.

BUG=b:269050063
TEST=Error message is printed if TCPC is configured to control Vconn.
BRANCH=none

Change-Id: I1777f0d307b1c835a39f2bdbe016dc908dcc8486
Signed-off-by: Kyle Lin <kylelinck@google.com>