zephyr/tests: emulator and tests for sm5803
This implements an emulator and some tests for the SM5803 charger. The
current tests do not exercise the driver as comprehensively as intended,
but those will be added in a later change.

One discovered bug in the driver is fixed, where some interrupts were
disabled immediately after being enabled.

BUG=b:242544165
TEST=twister -ciC -s drivers/drivers.sm5803
BRANCH=nissa

Change-Id: I9fc6abd26cd1b2e0cbc05b8e45cf94fc83abd08c
Signed-off-by: Peter Marheine <pmarheine@chromium.org>