Reduce AP communication failure spam
When communication with the AP is unavailable (not in S0, dead, etc.),
the EC spams these logs to the console:

    23-02-28 12:11:54.993 [53.641200 RMI: Unexpected command value in out bound message]
    23-02-28 12:11:55.004 [53.642300 STT: Failed to write SOC temp rv:1]
    23-02-28 12:11:55.008 [54.653200 RMI: Unexpected command value in out bound message]
    23-02-28 12:11:56.023 [54.654500 STT: Failed to write SOC temp rv:1]
    23-02-28 12:11:56.027 [55.663800 RMI: Unexpected command value in out bound message]
    23-02-28 12:11:57.042 [55.665500 STT: Failed to write SOC temp rv:1]

The actual content of the message is ignored by developers, and the
messages are instead interpreted to indicate that communication with
the AP is currently down. For nooglers and external developers, this
can lead to confusion and wasted time mis-attributing the log messages
to a different/separate error.

This CL removes the STT log message, and instead relies on
sb_rmi_mailbox_xfer() to log a single message when there is an AP
communication failure. The flag is reset when the AP communication is
restore.

BRANCH=none
BUG=b:271142544
TEST=tast run <IP> crash.WatchdogCrash, verify only a single error log
is output while the AP is rebooting

Change-Id: Iad08dbef64738d1a1f2d676fd47fbe7e66201e5e
Signed-off-by: Tim Van Patten <timvp@google.com>