system_safe_mode: Print process stack dump
Add the PLATFORM_EC_SYSTEM_SAFE_MODE_PRINT_STACK option for
printing the faulting process stack to the log. This must be done in
safe mode because the console buffer ignores log messages while in an
ISR.

This is a workaround until coredumps are fully supported.

This initial implementation only supports Cortex-M on Zephyr.

This implementation was barrowed from core/cortex-m/panic.c. It's
copied because the original code is not linked in Zephyr and some
small modifications were needed (ccprintf instead of panic_printf).

BUG=b:266084064
BRANCH=None
TEST=Observe stack dump on panic

Change-Id: Ied78ab7e6edca9cfa97c50323d94e39a3fca0eef
Signed-off-by: Rob Barnes <robbarnes@google.com>