zephyr: Provide all modules to toolchain implementations
Cmake support for the coreboot-sdk toolchain is located inside of the
EC module, however, the chameleon project does not want to load this
module.

There is nothing in the Zephyr build system that restricts
TOOLCHAIN_ROOT from existing inside of a module, so it's OK to pass
all modules to the toolchain providers.

BUG=b:271448278
BRANCH=none
TEST=chameleon build

Change-Id: I06523abfcb6e5f420c1a7520c6ec2750c2456f30
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>