zephyr/usba: define the correct symbol for port counts
Existing "dumb" code for USB-A port enables uses the USB_PORT_COUNT symbol
as the number of ports, whereas "smart" ports use
CONFIG_USB_PORT_POWER_SMART_PORT_COUNT.

Although only one of those will ever be set in a given configuration and
the downstream USB_PORT_ENABLE_COUNT value should always be correct, a
lot of older code uses one combination of an upstream value and the
downstream one depending on where the code lives and what configuration
it assumes. Fixing all users would be a large and difficult-to-make
change, so this change only makes the Zephyr behavior consistent with
ECOS.

One board depended on the dumb symbol being defined even when
USB-A control is disabled completely, and is modified so it
still builds.

BUG=b:271571241
TEST=zmake build -a; twister
BRANCH=nissa

Change-Id: I5509159d969107216c8297b1424ab583c0b370c0
Signed-off-by: Peter Marheine <pmarheine@chromium.org>