common: motion_sense: Call manage_activity only when needed
The global |sensor_active| which duplicates the power state is changed
in the CHIPSET_TASK.
It is used in lower priority tasks (MOTION_SENSE and HOOK), so can
change in a middle of a call.
It is usually fine, except when configuring activities, where we could
end up calling manage_activity() on a sensor that does not support it.

BUG=b:269225196
BRANCH=bugzzy,kukui (any branch where on body/off body is used).
TEST=Compile
Check buggzy ec.RW.dis to check the call to manage_activity() function
pointer is gone:
       228c            movs    r2, #140        ; 0x8c
       fb02 7005       mla     r0, r2, r5, r7
       2102            movs    r1, #2    <<< MOTIONSENSE_ACTIVITY_DOUBLE_TAP
       6c02            ldr     r2, [r0, #64]   ; 0x40 << sensor->driver
       4620            mov     r0, r4    <<< sensor
       6c15            ldr     r5, [r2, #64]   ; 0x40 << ..drive->manage_activity
       2201            movs    r2, #1    <<< 1
       47a8            blx     r5

Change-Id: I0060cfcb7b815cde54982e3e310b745cabed2306
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>