rex: fix parity on tcpc interrupts
the original tcpc interrupt pins was not configured to ACTIVE_LOW so all
logic involving the tcpc interrupt pins were reversed. This patch seeks
to move to Zephyr conventions for tcpc gpios.

gpio.dtsi is not edited due to it being an auto-generated file. It is
currently missing the ACTIVE_LOW flag which is instead set in
tcpc_config.

BUG=b:254148652
TEST=none
BRANCH=none
LOW_COVERAGE_REASON=Some untested code have been simplified, which does
                    not actually reduce code coverage even though it is
		    technically uncovered.

Change-Id: I5a4133a75d4eb089645633fa9b69fdb1d2e347b4
Signed-off-by: Jason Yuan <jasonyuan@google.com>