binman: Add CBI overlay section of fmap
The CBI section could be enabled if it's imported in the project.
The CBI section will be filled with 0xFF in binary. The size and
offset could be overridden with device tree config in chip or
program or project.

Add properties named 'preserve' and 'fill-byte' to flash layout
node to support new CBI section properties.

BRANCH=None
BUG=b:266972341
TEST=dump_fmap <ec.bin location>

Change-Id: I95cc4f1847a42e5086e78580d1e5f5fa7fd59dc9
Signed-off-by: Madhurima Paruchuri <mparuchuri@google.com>