binman: Add CBI overlay section of fmap
The CBI section could be enabled if it's imported in the project.
The CBI section will be filled with 0xFF in binary. The size and
offset could be overridden with device tree config in chip or
program or project.

A property named "preserve" is added, to determine the CBI portion
should not be updated during flashing of chip.

BRANCH=None
BUG=b:266972341
TEST=dump_fmap <ec.bin location>

Change-Id: I95cc4f1847a42e5086e78580d1e5f5fa7fd59dc9
Signed-off-by: Madhurima Paruchuri <mparuchuri@google.com>