bma4xx: avoid undefined results in ODR conversions
When small values were given to BMA4_ODR_TO_REG, the result would be
undefined because it would end up doing __fls(0) which is documented to
be undefined. Large values could also behave incorrectly, though the
result is defined. This change reimplements BMA4_ODR_TO_REG and
BMA4_REG_TO_ODR in a way that is rigorously proven to be correct and has
similar expected runtime cost.

This was discovered while implementing tests for the driver; the new
implementation passes the tests where the old one would modify
unintended bits of the ACCEL_CONFIG register if small values were passed
to set_data_rate.

BUG=b:256047664
TEST=bma4xx test suite now passes
BRANCH=none

Change-Id: I11764652e096881eb9118a25b032755405c165f6
Signed-off-by: Peter Marheine <pmarheine@chromium.org>