etool: Allow "ectool motionsense info" to dump all sensors
Previously if you wanted to dump all motionsense info then you needed
to go one-by-one through all of the sensors, calling ectool each
time. There's no reason why we can't just have the tool be able to
loop through all the sensors to make this easier. We'll make it so
that if a sensor number isn't provided that we'll just dump all of
them.

BUG=b:267680317
TEST=Run "ectool motionsense info"
BRANCH=none

Change-Id: Iebbe44cf20bca4470369147acebcd2ef237ca479
Signed-off-by: Douglas Anderson <dianders@chromium.org>