core/cortex-m: fix __data_lma_start when compiling with clang
The NPCX RO firmware image is composed of the 64-bytes NPCX image header
 (start at flash offset 0) + EC firmware image. After the EC chip boots
up, the EC booter copies only the EC image part from the flash to the
CODE RAM. As a result, the value of the __data_lma_start should be
assigned to the end of the VMA of the .rodata section
 (0x1008fcb0+0x4cb0=0x10094960) but the start of the LMA of .data
section (0x100949a0). Otherwise, it causes the init.s to copy the
incorrect Right-value to the global variable in the data section.

Sections:
Idx Name          Size      VMA       LMA       File off  Algn
  0 .header       00000040  10080000  10080000  00010000  2**0
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  1 .text         0000fcb0  10080000  10080040  00020000  2**2
                  CONTENTS, ALLOC, LOAD, READONLY, CODE
  2 .rodata       00004cb0  1008fcb0  1008fcf0  0002fcb0  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  3 .vtable       00000000  200c0000  200c0000  00034960  2**0
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  4 .bss          000029f8  200c0000  200c0000  00040000  2**3
                  ALLOC
  5 .data         000001a8  200c29f8  100949a0  000429f8  2**3
                  CONTENTS, ALLOC, LOAD, DATA

BRANCH=none
BUG=b:271333155
TEST=Verified the EC can run with clang build by David on the helipilot
project.

Change-Id: Ic9b6a2089792242f6a126f6b886536dde4dcd35d
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>