rex: split variant based on sensor inclusion
Add a new variant to the Rex BUILD.py that allows for skipping the
motion_sense task and all sensor related features. This requires:
1. Conditionally including the motionsense.dts file at the top level
   build.
2. Adding a new Kconfig overlay that disables the sensor related
   features for the sans-sensors build.
3. Factoring out the sensor related interrupts to only be included
   as an overlay when building the motionsense task.

These couldn't be done via a single Kconfig since Kconfig and DTS
processing are done at the same time (as far as we're concerned) in
CMake. This means we can't conditionally include the DTS files at
configuration time and instead require a new overlay.

BRANCH=none
BUG=b:272529568
TEST=zmake build rex rex-sans-sensors

Change-Id: I23df21f41f8a9cd4336187d04034e7e9de10110d
Signed-off-by: Yuval Peress <peress@google.com>