ec: Add async FLASH_PROTECT
Create an async FLASH_PROTECT command that is enabled when
CONFIG_FLASH_PROTECT_DEFERRED is defined. Disable this configuration
for boards cerise, damu, and willow where there is no flash memory
available. Disable this configuration for board hyperdebug where the
deferred implementation in not possible because CONFIG_COMMON_RUNTIME is
undefined. Enable this configuration otherwise in both the ec and zephyr
configuration files.

BRANCH=nocturne
BUG=b:116396469
TEST=make buildall -j

Change-Id: I0605cc600841ec2ec3c1dd2d3d3598b28d20ffcb
Signed-off-by: Firas Sammoura <fsammoura@google.com>