ec: Add async FLASH_PROTECT
Create an async FLASH_PROTECT command that is enabled when
CONFIG_FLASH_PROTECT_DEFERRED is defined. Disable this configuration
for boards cerise, damu, and willow where there is no flash room
available. Disable this configuration for board hyperdebug where the
deferred implenetation in not possible because CONFIG_COMMON_RUNTIME is
undefined.

BRANCH=nocturne
BUG=b:116396469
TEST=make buildall -j

Change-Id: I0605cc600841ec2ec3c1dd2d3d3598b28d20ffcb
Signed-off-by: Firas Sammoura <fsammoura@google.com>