npcx9: Move RW region so we can write protect RO region
The npcx9 internal flash write protect scheme is very limited. The
smallest protectable region is half of the entire flash device, so
allocate RO and RW regions to match the hardware.

Note: The EC image in the BIOS must match this new region allocation.
The only way to do this is to manually update the BIOS and EC. When the
BIOS triggers a software sync with mismatched region configurations, the
system will endlessly try to software sync.

BRANCH=none
BUG=b:260815079
TEST=FAFT firmware_ECCbiEeprom passes

Change-Id: Id5b9b2bc8d7858f3cb90ef39300ef1f654a080e6
Signed-off-by: Caveh Jalali <caveh@chromium.org>