EFS2: Hard reset after shutdown in recovery mode
Currently, when a system shuts down in recovery mode, the EC jumps to
RW. This additional entry point to the RW increases the dependency on
the RO.

This patch changes sysjump to a hard reset. This change makes an RW be
only entered from the RO's reset vector. After this reset, the AP is set
to AP_OFF.

We assume AP wants to reboot instead of shut down the system. If AP
wants to keep the system off after shutdown, it's supposed to issue
SYSTEM_RESET_LEAVE_AP_OFF with FLAG_ON_AP_SHUTDOWN beforehand.

BUG=b:201099749, b:294765831, b:263424688
TEST=Eldrid. Shutdown from recovery screen and dev screen from menu and
     by power button.

Cq-Depend: chromium:4775107
Change-Id: I9add38fdf8925f143c58e705b5070a782f5d2cee