efs2: Hard reset after shutdown in recovery mode
Currently, when a system shuts down in recovery mode, the EC jumps to
RW. This additional entry point to the RW increases the dependency on
the RO.

This patch changes sysjump to a hard reset. This change makes an RW be
only entered from the RO's reset vector. After this reset, the AP is set
to AP_OFF.

BUG=b:201099749
BRANCH=None
TEST=Agah. Shutdown from recovery screen.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I9add38fdf8925f143c58e705b5070a782f5d2cee