EFS2: Hard reset after shutdown in recovery mode
Currently, when a system shuts down in recovery mode, the EC jumps to
an RW. This additional entry point increases RW's dependency on the RO
because what the RO does (after vboot_main) affects RW's run.

This patch removes a sysjump in a shutdown hook and instead schedules a
hard reset on entry to recovery mode. The net effect is RW can be only
entered from vboot_main. It makes an RW instance more independent of the
RO.

We assume the AP wants to reboot the system. If the AP wants to shutdown
and keep the system off, it needs to schedule SYSTEM_RESET_LEAVE_AP_OFF
on shutdown.

BUG=b:201099749, b:294765831, b:263424688
TEST=Eldrid. Shutdown from recovery screen and dev screen from menu and
     by power button. Verify the DUT is in RO & G3.
TEST=Vell. Unplug an external disk after recovery is finished. Observe
     the DUT successfully reboots. Verify the EC receives a sysjump
     command and the AP warm-resets.
TEST=Vell. From the recovery FW screen, close the lid. Observe the EC
     executes the SHUTDOWN_COMPLETE hook. Observe the DUT reboots in
     normal mode but Depthcharge turns off the DUT immediately.
TEST=Vell. Run apshutdown in recovery mode. Observe DUT stays off.
TEST=Vell. Run apshutdown -d in recovery mode. Observe DUT reboots.
TEST=Vell. Run 'ectool reboot-ec cold at-shutdown' then 'ectool
     batterycutoff at-shutdown'. Observe cutoff succeeds after shutdown.

Cq-Depend: chromium:4775107
Change-Id: I9add38fdf8925f143c58e705b5070a782f5d2cee